Use a bundler rather than cram all into one file
