+++ product_chooser.js
@@ -1,150 +1,4 @@
-const fields = {
-  "ring-plain": {
-    field: "ring_type",
-    value: "plain-shaped"
-  },
-  "ring-diamond": {
-    field: "ring_type",
-    value: "gem-set"
-  },
-  "carat-9": {
-    field: "carat",
-    value: "9ct",
-    excl_attrs: {
-      "metal": ["titanium", "platinum"]
-    },
-    excl_fields: [
-      "metal-titanium", "metal-platinum"
-    ]
-  },
-  "carat-18": {
-    field: "carat",
-    value: "18ct",
-    excl_attrs: {
-      "metal": ["titanium", "platinum"]
-    },
-    excl_fields: [
-      "metal-titanium",
-      "metal-platinum",
-    ]
-  },
-  "style-flat": {
-    field: "style",
-    value: "flat",
-  },
-  "style-easyfit": {
-    field: "style",
-    value: "easy-fit",
-  },
-  "style-court": {
-    field: "style",
-    value: "court",
-  },
-  "style-dshape": {
-    field: "style",
-    value: "d-shape",
-  },
-  "width-2": {
-    field: "width",
-    value: "2mm",
-    excl_attrs: {
-      "metal": ["titanium"]
-    },
-    excl_fields: [
-      "metal-titanium"
-    ]
-  },
-  "width-2.5": {
-    field: "width",
-    value: "2-5mm",
-    excl_attrs: {
-      "metal": ["rose-gold", "titanium"]
-    },
-    excl_fields: [
-      "metal-titanium",
-      "metal-rosegold",
-    ]
-  },
-  "width-3": {
-    field: "width",
-    value: "3mm",
-    excl_attrs: {
-      "metal": ["titanium"]
-    },
-    excl_fields: [
-      "metal-titanium"
-    ]
-  },
-  "width-4": {
-    field: "width",
-    value: "4mm",
-  },
-  "width-5": {
-    field: "width",
-    value: "5mm",
-  },
-  "width-6": {
-    field: "width",
-    value: "6mm",
-    excl_attrs: {
-      "metal": ["rose-gold"]
-    },
-    excl_fields: [
-      "metal-rosegold"
-    ]
-  },
-  "width-8": {
-    field: "width",
-    value: "8mm",
-    excl_attrs: {
-      "metal": ["rose-gold"]
-    },
-    excl_fields: [
-      "metal-rosegold"
-    ]
-  },
-  "metal-yellowgold": {
-    field: "metal",
-    value: "yellow-gold"
-  },
-  "metal-rosegold": {
-    field: "metal",
-    value: "rose-gold",
-    excl_attrs: {
-      "width": ["6mm", "8mm", "2.5mm"]
-    },
-    excl_fields: [
-      "width-6",
-      "width-8",
-      "width-2.5"
-    ]
-  },
-  "metal-whitegold": {
-    field: "metal",
-    value: "white-gold"
-  },
-  "metal-platinum": {
-    field: "metal",
-    value: "platinum",
-    excl_attrs: {
-      "carat": ["9ct", "18ct"]
-    },
-    excl_fields: [
-      "carat-9", "carat-18"
-    ]
-  },
-  "metal-titanium": {
-    field: "metal",
-    value: "titanium",
-    excl_attrs: {
-      "width": ["2mm", "2.5mm", "3mm"],
-      "carat": ["9ct", "18ct"]
-    },
-    excl_fields: [
-      "carat-9", "carat-18", "width-2", "width-2.5", "width-3"
-    ]
-  }
-};
+import { fields } from './product_fields';
 
 // Object representing the current state of selections
 // the user has made
@@ -298,9 +152,7 @@ const getImageSrc = (elem) => {
     affected_field === "width" ? elem_details.value : (fields[selections.width]?.value || "4mm")
   ];
 
-  let key = selected.join("-");
-
-  return "/wp-content/uploads/2020/08/" + key + ".jpg"
+  return imageUrl(selected.join("-"));
 };
 
 // For a given element, fetch its child image and update the src attribute
@@ -344,7 +196,9 @@ const skipToNextSection = (event) => {
 }
 
 const initialHide = () => {
-  Object.keys(selections).forEach((pfx) => {
+  let sects = Object.keys(selections);
+  sects.push("spec");
+  sects.forEach((pfx) => {
     if (pfx === "metal") {
       return
     };
@@ -354,6 +208,37 @@ const initialHide = () => {
       section.style.display = "none";
     }
   });
+}
+
+const setSpecText = (carat, metal, style, width) => {
+  const target = document.getElementById('spec-text');
+  target.innerHTML = `Your chosen ring is a ${carat ? carat + ' ' : ''}${metal} ${style} wedding ring with a finger width of ${width}.`
+}
+
+const setSpecImage = (metal, style, width) => {
+  const target = document.getElementById('spec-image');
+  let selected = [
+    style || "court",
+    metal || "yellow-gold",
+    width || "4mm"
+  ];
+
+  target.src = imageUrl(selected.join("-"));
+  target.srcset = "";
+}
+
+const imageUrl = (key) => {
+  return "/wp-content/uploads/2020/08/" + key + ".jpg";
+}
+
+const updateSpec = () => {
+  const carat = fields[selections.carat]?.value;
+  const metal = fields[selections.metal]?.value;
+  const style = fields[selections.style]?.value;
+  const width = fields[selections.width]?.value;
+
+  setSpecText(carat, metal, style, width);
+  setSpecImage(metal, style, width);
 }
 
 // Initially, the URL is invalid and so we just link to '#'.
